Add AuditInfoTest
